:up: chkimgの指定スコア範囲を80-99に変更

+++ evaluator/src/error.rs
@@ -536,7 +536,7 @@ pub enum UErrorMessage {
     InvalidErrorLine(usize),
     FailedToGetObject,
     GdiError(String),
-    GivenNumberIsOutOfRange(f64, f64),
+    GivenNumberIsOutOfRange(i32, i32),
     WebSocketTimeout(Value),
     WebSocketConnectionError(String),
     InvalidParamType(String, ParamTypeDetail),
